+++ main/components/bloomfilter/bloom.h
@@ -31,13 +31,13 @@ typedef struct bloom
   uint16_t bits;          // number of bits in bloom filter
   uint8_t bytes;          // number of bytes in bloom filter
   uint16_t hashes;
-  double error;           // error rate of false positives
+  float error;           // error rate of false positives
 
   // Fields below are private to the implementation. These may go away or
   // change incompatibly at any moment. Client code MUST NOT access or rely
   // on these.
   uint8_t ready;
-  double bpe;             // bits per element
+  float bpe;             // bits per element
   uint8_t * bf;
 } __attribute__((packed)) bloom_t;
 
@@ -69,7 +69,7 @@ typedef struct bloom
  *     1 - on failure
  *
  */
-int8_t bloom_init2(struct bloom * bloom, uint8_t entries, double error);
+int8_t bloom_init2(struct bloom * bloom, uint8_t entries, float error);
 
 
 /** ***************************************************************************
